Add a custom generator to efficiently find target pages.
Rather than fetching each file page and parsing it to find the
Geograph ID, take advangtage of the fact that the sort key prefix in
[[:Category:Images from the Geograph British Isles project]] is the
Geograph ID, so we can iterate over this category fetching just page
titles and sort key prefixes, and quickly skip over images for which
no high-res version exists on Geograph.

This is implemented as an optimisation: any checks done by the
generator are repeated on any file that it yields.  Indeed, if it has
any problems dealing with a file (e.g. the sort key is unparseable),
it passes that file on to the main processor for proper error
reporting.

This makes the bot much faster at skipping files.  In a quick test, it
processed past Geograph ID 10000 in 37 seconds, of which ten seconds
was waiting for database lag to subside.