Compare aspect ratios of images before uploading.
If the aspect ratios of the high- and low-resolution images in
gridimage_size differ by more than 1%, treat the image as ineligible.
That avoids some really obvious cases where the high-res picture is
significantly different from the low-res one.  To cope with cases where
the high-res image is rotate, also check the inverse of the aspect
ratio.

This change has been in production since Friday.