fix_locations: don't add credit line if title has changed on Geograph.
The plan here is to only add a credit line it the title is the same as
it was when GeographBot uploaded the picture.  This is a bit
unreliable because GeographBot didn't mark the end of the title, but
this will only be a problem for titles that are prefixes of the
original one.