location: Add "region" attribute to geocoding templates.
Some sources say it's obsolete, but GeoHack still uses it, so it's
worth adding.  Working out the correct one is mildly tricky for the GB
grid and difficult for the Irish Grid, so I resort to using MapIt.
This might end up costing rather a lot in MapIt API calls, so it
should probably be modified to tolerate failure of those.

Currently, we return:

region:GB-GBN for points in Great Britain
region:GB-NIR for points in Northern Ireland
region:IE for points in Ireland
region:IM for points on the Isle of Man