Take advantage of prop=categories to simplify upgrade_size
With prop=categories&clprop=sortkey, I can get the Geograph IDs for
files from any generator.  This is a bit indirect, but it saves having
to run two generators in parallel to get imageinfo and sortkey, so I
like it.

Weirdly, cllimit (the maximum number of categories to return) seems to
apply to the whole query, so set it to "max".  iilimit, on the other
hand, is per-file.