Add links to edit summaries
These links are intended to point (via redirects) to pages describing
the bot tasks whose edit summaries they appear in.  This is a
requirement of the bot policy on enwiki and while it isn't required on
Commons it does seem like a good idea.

My approach has been to attach the links to existing text in the edit
summary (usually the main verb) so as not to make the visible text
gratuitously longer.