Add geocoding statements if we're updating SDC anyway
Specifically if we're updating one of the two geocoding statements and
the other one is missing then we'll add it if Geograph has suitable
co-ordinates.  Now that we can tag statements with sources it doesn't
seem dangerous to add them, and if we're already editing structured data
then this won't fill up people's watchlists any more than we already
are.