Add credit lines even if the title might have changed on Geograph
I can't see any harm in adding the current title even if it isn't the
one that was current when the file was uploaded.  I can't work out why
I included that check in the first place, and as far as I can tell the
code doesn't do what the documentation says it does in that it doesn't
special-case GeographBot uploads.

My best guess is that I was worried about the case where a Geograph
file had been overwritten and the new version was a big enough change
that the person editing it had their own copyright in it and they were
only willing to licence it under the old title and not the new one.
But that seems quite an unlikely set of circumstances (and would
almost certainly by a breach of COM:OVERWRITE), so I don't think it's
anywhere near as important as files that don't have a correct title at
all.